Home page 구조 분리
