+++ todo-list/src/pages/home.jsx
@@ -3,13 +3,14 @@ import { useParams } from "react-router-dom";
 import styled from "styled-components";
 import TodoItem from "../components/TodoItem";
 import Calendar from "../components/Calendar";
+import TodoList_Section from "../components/TodoList_Section";
+import AddTodo_Section from "../components/AddTodo_Section";
 
 import {
   BaseContainer,
   BaseInnerContainer,
-  Ellipse,
   GridLayout,
-} from "../components/styledComponents";
+} from "../styles/styledComponents";
 
 const Home = () => {
   const { id } = useParams();
@@ -29,53 +30,10 @@ const Home = () => {
             </CalendarInnerContainer>
           </CalendarContainer>
           <TODOContainer>
-            <TODOInnerContainer>
-              <ContainerTitle>ADD TODO</ContainerTitle>
-              <TODOInput placeholder="할 일을 추가하세요."></TODOInput>
-              <ButtonWrapper>
-                <AddButton>
-                  <AddButtonImg src="../public/Done.png"></AddButtonImg>
-                </AddButton>
-              </ButtonWrapper>
-            </TODOInnerContainer>
+            <AddTodo_Section></AddTodo_Section>
           </TODOContainer>
           <ListContainer>
-            <ListInnerContainer>
-              <ContainerTitle>MY TODO LIST</ContainerTitle>
-              <TodoListWrapper>
-                {/* 실제 db에서 데이터 가져와야 함: TODO LIST */}
-                <TodoItem
-                  id={0}
-                  isChecked={true}
-                  content={"할 일 1"}
-                  emoticon={"none"}
-                ></TodoItem>
-                <TodoItem
-                  id={0}
-                  isChecked={true}
-                  content={"할 일 2"}
-                  emoticon={"none"}
-                ></TodoItem>
-                <TodoItem
-                  id={0}
-                  isChecked={true}
-                  content={"할 일 3"}
-                  emoticon={"none"}
-                ></TodoItem>
-                <TodoItem
-                  id={0}
-                  isChecked={true}
-                  content={"할 일 4"}
-                  emoticon={"none"}
-                ></TodoItem>
-                <TodoItem
-                  id={0}
-                  isChecked={true}
-                  content={"할 일 5"}
-                  emoticon={"none"}
-                ></TodoItem>
-              </TodoListWrapper>
-            </ListInnerContainer>
+            <TodoList_Section></TodoList_Section>
           </ListContainer>
         </GridLayout>
       </ContentWrapper>
@@ -130,98 +88,11 @@ const TODOContainer = styled(BaseContainer)`
   grid-area: 1 / 2 / 2 / 4; /* 1행 2~3열 */
 `;
 
-const TODOInnerContainer = styled(BaseInnerContainer)`
-  width: 653px;
-  height: 249px;
-  flex-shrink: 0;
-`;
-
 const ListContainer = styled(BaseContainer)`
   width: 1059px;
   height: 261px;
   flex-shrink: 0;
   grid-area: 2 / 1 / 3 / 4; /* 2행 전체 */
-`;
-
-const ListInnerContainer = styled(BaseInnerContainer)`
-  width: 1047px;
-  height: 249px;
-  flex-shrink: 0;
-`;
-
-const ContainerTitle = styled.div`
-  margin-left: 30px;
-  margin-top: 5px;
-  width: 100%;
-  padding: 5px;
-  height: auto;
-  color: #000;
-  text-align: left;
-  font-family: Grandstander;
-  font-size: 1.5rem;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-  background: transparent;
-`;
-
-const TODOInput = styled.textarea`
-  width: 90%;
-  height: 140px;
-  flex-shrink: 0;
-  border-radius: 7px;
-  border: 1px solid #c8c8c8;
-  background: #fff;
-  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-  padding: 10px;
-  font-size: 15px;
-  color: #6f6f6f;
-  font-family: Pretendard;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  resize: none;
-`;
-
-const ButtonWrapper = styled.div`
-  margin-right: 40px;
-  margin-bottom: 5px;
-  width: 100%;
-  padding: 10px;
-  height: auto;
-  background: transparent;
-  display: flex;
-  justify-content: end;
-`;
-
-const AddButton = styled.div`
-  border-radius: 10px;
-  border: 1px solid #7c7c7c;
-  background: #fff;
-  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-  cursor: pointer;
-  width: 35px;
-  height: 27px;
-  flex-shrink: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const AddButtonImg = styled.img`
-  width: 20px;
-  height: 20px;
-  flex-shrink: 0;
-  background: transparent;
-`;
-
-const TodoListWrapper = styled.div`
-  width: 100%;
-  height: 100%;
-  margin: 5px 0px 1px 0px;
-  overflow-y: scroll;
-  overflow-x: hidden;
-  background: transparent;
 `;
 
 export default Home;
